Created util file with common utilities
